==> notebooks/01_introduction.Rmd <==
---
title: "Introduction"
author: "Benzon Carlitos Salazar"
date: "October 22, 2022"
---

# Introduction

My goal for this project is to exercise my data analysis skills. I have chosen 
to work with the Drug Reviews dataset [@Gräßer2018] after spending some time in 
the UCI Machine Learning Repository [@Dua2019].

I will focus on some illustrative data visualizations, and I will also try to 
stay true to the original paper and do a very basic, beginner-friendly sentiment 
analysis of drug satisfaction.

## Load and check the data

```{r setup}
library(magrittr)

train <- readr::read_tsv(here::here("data-raw", "drugsComTrain_raw.tsv"))
test <- readr::read_tsv(here::here("data-raw", "drugsComTest_raw.tsv"))

full <- dplyr::bind_rows(train, test)

# Check our data
str(full)

colnames(full)
```

Based on our output, we get `r ncol(full)` variables (or columns) and 
`r dplyr::count(full)` observations (or rows). That's a pretty lengthy dataset!

Our data dictionary is as follows:

| Variable Name | Description                             |
|:------------- |:--------------------------------------- |
| ...1          | Identifier                              |
| drugName      | Name of drug                            | 
| condition     | Name of condition                       |
| review        | Patient review                          |
| rating        | 10 start patient rating                 |
| date          | Date of review entry                    |
| usefulCount   | Number of users who found review useful |

## Tokenization and data cleaning

The first column, `...1` looks to be an identifier, so we rename that column as 
and `id`, and then we do some minor tokenization to enforce consistency. 
Additionally, we will utilize the `lubridate` package to break down the date 
column to census month and census year, to see if it will give us some additional 
insights.

We will also remove any special characters in the `drugName` and `condition` 
columns, turn white spaces into underscores, while keeping the original text for 
the `review` for now.

```{r tokenization}
full <-
  full %>%
  dplyr::rename(id = "...1")

full$drugName <- stringr::str_replace_all(full$drugName, "[^[:alnum:]]", "_")
full$condition <- stringr::str_replace_all(full$condition, "[^[:alnum:]]", "_")
full$drugName <- tolower(full$drugName)
full$condition <- tolower(full$condition)
full$date <- lubridate::mdy(full$date)
full$rating <- as.numeric(full$rating)
full$usefulCount <- as.numeric(full$usefulCount)

full <-
  full %>%
  dplyr::mutate(census_month = lubridate::month(date)) %>%
  dplyr::mutate(census_year = lubridate::year(date)) %>%
  dplyr::rename(full_date = date)
```
