==> notebooks/02_summary_statistics.Rmd <==
---
title: "Summary statistics"
author: "Benzon Carlitos Salazar"
date: "`r format(Sys.time(), '%B %d, %Y')`"
---

```{r setup, warning=FALSE, include=FALSE, echo=FALSE}
library(magrittr)

full <- readRDS(here::here("data", "full.rds"))
```

# Summary statistics

Doing some summarizations, we can see count distributions for each column. This 
will also help us determine if there exists any NA columns.

## Drugs

```{r drug_summarization}
drug_count <-
  full %>%
  dplyr::select(drugName) %>%
  dplyr::group_by(drugName) %>%
  dplyr::summarise(count = dplyr::n())
```

What is the top 10 most used drug?

```{r top_10_most_drug}
drug_count %>%
dplyr::arrange(desc(count)) %>%
dplyr::slice_head(n = 10) %>%
knitr::kable()
```

What is the top 10 least used drug?

```{r top_10_least_drug}
drug_count %>%
dplyr::arrange(count) %>%
dplyr::slice_head(n = 10) %>%
knitr::kable()
```

The top 10 least used drug is actually misleading, as there are 
`r drug_count %>% dplyr::filter(count == 1) %>% dplyr::count() %>% as.integer` 
drugs that all have a count of 1.

Are there any NAs or NULLs for `drugName`?

```{r na_nulls}
drug_count %>%
dplyr::filter(is.na(drugName) | is.null(drugName)) %>%
dplyr::select(count) %>%
knitr::kable()
```

## Patient Condition

```{r condition_summarization}
condition_count <-
  full %>%
  dplyr::select(condition) %>%
  dplyr::group_by(condition) %>%
  dplyr::summarise(count = dplyr::n())
```

What are the top 10 patient conditions?

```{r pat_condition_top_10}
condition_count %>%
dplyr::arrange(desc(count)) %>%
dplyr::slice_head(n = 10) %>%
knitr::kable()
```

What are the least patient conditions?

```{r pat_condition_least}
condition_count %>%
dplyr::arrange(count) %>%
dplyr::slice_head(n = 35) %>%
knitr::kable()
```

I am unsure what to make of the `*_span_users_found_this_comment_helpful_`, so 
we will be removing those later. But ignoring that weirdness, we can see that 
`amyotrophic_lateral_sclerosis` is the least mentioned condition, followed by 
`anti_nmda_receptor_encephalitis`.

Are there any NAs or NULLs for patient `condition`?

```{r na_nulls}
na_condition <-
  condition_count %>%
  dplyr::filter(is.na(condition) | is.null(condition)) %>%
  dplyr::select(count)
```

There appears to be `r as.integer(na_condition)` NA or NULL conditions. Since 
there are so many, I have decided to leave them in and just convert each NA/NULL 
conditions to `unknown_condition` value.

```{r replace_na}
full$condition <- tidyr::replace_na(full$condition, "unknown_condition")

# Check
full %>%
dplyr::filter(condition == "unknown_condition") %>%
dplyr::count() %>%
knitr::kable()
```

## Ratings

Since the ratings are 1-10, we will categorize them into 3 levels:

* low_rating = 0-3
* mid_rating = 4-7
* high_rating = 8-10

First, we create our row levels, then we add a new `drug_rating` column.

```{r rating_row_lvls}
rating_row_lvl <- 
  c("Low [0-3]", 
    "Mid [4-7]", 
    "High [8-10]", 
    "Unknown")

full_proc_rating <-
  full %>%
  dplyr::mutate(
    drug_rating = factor(
      dplyr::case_when(
        rating > 0 & rating <= 3 ~ rating_row_lvl[1],
        rating >= 4 & rating <= 7 ~ rating_row_lvl[2],
        rating >= 8 & rating <= 10 ~ rating_row_lvl[3],
        TRUE ~ rating_row_lvl[4]
      ), levels = rating_row_lvl))
```

Then, we utilize `{gtsummary}` package to see count distribution and frequency 
distributions.

```{r rating_gt}
full_proc_rating %>%
dplyr::select(drug_rating) %>%
gtsummary::tbl_summary(label = list(drug_rating ~ "Drug Rating")) %>%
gtsummary::bold_labels() %>%
gtsummary::modify_footnote(update = dplyr::everything() ~ NA)
```

Additionally, we can plot it as well.

```{r rating_factor_plt, fig.align = "center"}
full_proc_rating %>%
ggplot2::ggplot(ggplot2::aes(x = drug_rating)) +
ggplot2::geom_bar(fill = "#0099f9") + 
ggplot2::theme_classic() +
ggplot2::scale_y_continuous(
  labels = scales::label_number(suffix = "K", scale = 1e-3)
) +
ggplot2::labs(
  title = "Patient rating distribution",
  x = "Drug Rating Categories",
  y = "Patient Count"
)
```

We see that the general consensus leans towards a high rating for each drugs. 
Let's drill down further to see what is the highest overall rating.

```{r rating_drill_down}
full %>%
dplyr::mutate(rating__char = as.character(rating)) %>%
dplyr::select(rating__char) %>%
gtsummary::tbl_summary(label = list(rating__char ~ "Drug Rating")) %>%
gtsummary::bold_labels() %>%
gtsummary::modify_footnote(update = dplyr::everything() ~ NA)
```

Let's plot our summaries.

```{r rating_plots, fig.align = "center"}
full %>%
ggplot2::ggplot(ggplot2::aes(x = rating)) + 
ggplot2::geom_bar(fill = "#0099f9") +
ggplot2::theme_classic() +
ggplot2::labs(
  title = "Patient rating distribution",
  x = "Drug Ratings",
  y = "Patient Count"
)

rating__sums <-
  full %>%
  dplyr::mutate(rating__char = as.character(rating)) %>%
  dplyr::group_by(rating__char) %>%
  dplyr::summarise(count = dplyr::n())

max_tmp_cnt <- max(rating__sums$count)
max_tmp_rating <-
  rating__sums %>%
  dplyr::filter(count == max_tmp_cnt) %>%
  dplyr::select(rating__char)
```

We see that patients mostly rated their drugs `r as.integer(max_tmp_rating)`, 
with a maximum count of `r as.integer(max_tmp_cnt)`. This lets us know that the 
attitude towards each patient's drug is positive.

## Review Dates

Using the `census_month` and `census_year`, we should be able to see count 
distributions per month and per year. To make our analysis easier, we will 
summarize the census month and year separately.

First, we convert each month numerical representation to their correct string 
representation, and then summarize with `{gtsummary}`.

```{r review_months, fig.align = "center"}
review_census_month <-
  full %>%
  dplyr::mutate(
    census_month__abbr = factor(month.abb[census_month], levels = month.abb)
  )

review_census_month %>%
dplyr::select(census_month__abbr) %>%
gtsummary::tbl_summary(
  label = list(
    census_month__abbr ~ "Census Month of Review"
  )
) %>%
gtsummary::bold_labels() %>%
gtsummary::modify_footnote(update = dplyr::everything() ~ NA)

review_census_month %>%
ggplot2::ggplot(ggplot2::aes(x = census_month__abbr)) + 
ggplot2::geom_bar(fill = "#0099f9") + 
ggplot2::theme_classic() +
ggplot2::scale_y_continuous(
  labels = scales::label_number(suffix = "K", scale = 1e-3)
) +
ggplot2::labs(
  title = "Patient Review per Month",
  x = "Census Month",
  y = "Patient Count"
)
```

```{r tmp_month, include=FALSE, echo=FALSE, warning=FALSE}
tmp__month <-
  review_census_month %>%
  dplyr::group_by(census_month__abbr) %>%
  dplyr::summarise(count = dplyr::n())

max_count <- max(tmp__month$count)

max_month <-
  tmp__month %>%
  dplyr::filter(count == max_count) %>%
  dplyr::select(census_month__abbr)
```

Based on the summary table, reviews look pretty evenly distributed throughout 
the months. Although, it looks like `r as.integer(max_month)` has the most 
reviews, with `r as.integer(max_count)` review counts.

Next, we take a look at the census year.

```{r review_years, fig.align = "center"}
full %>%
dplyr::select(census_year) %>%
gtsummary::tbl_summary(
  label = list(census_year ~ "Census Year of Review")
) %>%
gtsummary::bold_labels() %>%
gtsummary::modify_footnote(update = dplyr::everything() ~ NA)

full %>%
dplyr::select(census_year) %>%
ggplot2::ggplot(ggplot2::aes(x = census_year)) + 
ggplot2::theme_classic() +
ggplot2::scale_y_continuous(
  labels = scales::label_number(suffix = "K", scale = 1e-3)
) +
ggplot2::geom_bar(fill = "#0099f9") + 
ggplot2::labs(
  title = "Patient Review per Year",
  x = "Census Year",
  y = "Patient Count"
)
```

```{r tmp_year, include=FALSE, echo=FALSE, warning=FALSE}
tmp__yr <-
  full %>%
  dplyr::group_by(census_year) %>%
  dplyr::summarise(count = dplyr::n())

max_year_cnt <- max(tmp__yr$count)
max_year <- 
  tmp__yr %>%
  dplyr::filter(count == max_year_cnt) %>%
  dplyr::select(census_year)
```

At first glance, it looks like we see a steady increase in the number of patients 
review. We especially see a jump from 2014 to 2015, with year 
`r as.integer(max_year)` having the most reviews of `r max_year_cnt`.

### Census year and month

Taking a look at the census year and months together, we can see which month for 
each year the distribution of patient reviews.

Let's first look at the trend of reviews.

```{r review_trends, fig.align = "center"}
census_month_year <-
  full %>%
  dplyr::select(census_month, census_year) %>%
  dplyr::mutate(
    census_month = factor(month.abb[census_month], levels = month.abb))

census_month_year %>%
dplyr::group_by(census_month, census_year) %>%
dplyr::summarise(rev_count = dplyr::n(), .groups = "keep") %>%
dplyr::arrange(census_year, census_month) %>%
dplyr::ungroup() %>%
dplyr::mutate(time = dplyr::row_number()) %>%
ggplot2::ggplot(ggplot2::aes(x = time, y = rev_count)) +
ggplot2::geom_point(shape = 21, color = "black", fill = "#69b3a2", size = 2) + 
ggplot2::geom_line(color = "black") +
ggplot2::theme_classic() +
ggplot2::scale_y_continuous(
  labels = scales::label_number(suffix = "K", scale = 1e-3)
) +
ggplot2::labs(
  title = "Patient Review count overtime",
  x = "Time period (t)",
  y = "Review Count",
)
```

We see a couple of peaks and valleys here and there. At around t=23, we see the 
first peak, but we can also see that by t=25, we get the biggest drop in patient 
responses. Then the biggest climb from t=80 up to t=90 (approximately).

Let's look at some summaries.

```{r census_month_and_year, fig.align = "center"}
census_month_year %>%
gtsummary::tbl_summary(
  by = census_year,
  label = list(census_month ~ "Census Month")
) %>%
gtsummary::bold_labels() %>%
gtsummary::modify_footnote(update = dplyr::everything() ~ NA)

census_month_year %>%
dplyr::group_by(census_month, census_year) %>%
dplyr::summarise(rev_count = dplyr::n(), .groups = "keep") %>%
ggplot2::ggplot(
  ggplot2::aes(x = census_year, y = census_month, fill = rev_count)
) +
ggplot2::geom_tile() +
ggplot2::theme_classic() +
ggplot2::scale_fill_gradient(low="white", high="blue") +
ggplot2::labs(
  title = "Patient Review per Month per Year",
  x = "Census Year",
  y = "Census Month",
  fill = "Review Count"
)
```

The heatmap provides a great visualization of when we see an increase in patient 
responses. From March of 2015 all the way to May of 2017, we see strong numbers 
of responses. Then, at about April of 2017, we see some decline all the way to 
November of that same year.

## Key Takeaways

Excluding the `usefulCount` column (I would like to clump that together with the 
patient reviews), and outside the sentiment of the reviews, we can create the 
following conclusions:

* Levonorgestrel, a progestin, is the most reviewed drug, followed by 
Etonogestrel.
* Most patients rated their drug pretty high, with mostly ratings of 8, 9, 
and 10. With 32% of all ratings being a 10/10.
* Reviews are happening consistently every month.
* We see an increase in reviews, however, in years 2015, 2016, and 2017, with 
2016 being the year with most reviews.

```{r save_and_clean, include=FALSE, echo=FALSE, warning=FALSE}
saveRDS(full, here::here("data", "full.rds"))
rm(list = ls())
gc()
```
